Schedule pills

+++ app/src/main/java/com/example/medregister/RegisterPillsActivity.java
@@ -36,6 +36,7 @@ public class RegisterPillsActivity extends AppCompatActivity {
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        //inflate view
         setContentView(R.layout.activity_registerpills);
         Log.d(TAG, "onCreate: started.");
         setTitle(R.string.register_pills_title);
@@ -140,13 +141,11 @@ public class RegisterPillsActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.delete_all_pills:
-                pillViewModel.deleteAllPills();
-                Toast.makeText(this, R.string.all_pills_deleted, Toast.LENGTH_SHORT).show();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
+        if (item.getItemId() == R.id.delete_all_pills) {
+            pillViewModel.deleteAllPills();
+            Toast.makeText(this, R.string.all_pills_deleted, Toast.LENGTH_SHORT).show();
+            return true;
         }
+        return super.onOptionsItemSelected(item);
     }
 }
